+++ src/plugin/regions/index.js
@@ -13,6 +13,7 @@
  * @property {number[]} maxRegions Maximum number of regions that may be created by the user at one time.
  * `initPlugin('regions')`
  * @property {function} formatTimeCallback Allows custom formating for region tooltip.
+ * @property {?number} edgeScrollWidth='5% from container edges' Optional width for edgeScroll to start
  */
 
 /**
@@ -29,6 +30,7 @@
  * @property {?number} channelIdx Select channel to draw the region on (if there are multiple channel waveforms).
  * @property {?object} handleStyle A set of CSS properties used to style the left and right handle.
  * @property {?boolean} preventContextMenu=false Determines whether the context menu is prevented from being opened.
+ * @property {boolean} showTooltip=true Enable/disable tooltip displaying start and end times when hovering over region.
  */
 
 import {Region} from "./region.js";
@@ -115,6 +117,7 @@ export default class RegionsPlugin {
             }
         };
         this.maxRegions = params.maxRegions;
+        this.regionsMinLength = params.regionsMinLength || null;
 
         // turn the plugin instance into an observer
         const observerPrototypeKeys = Object.getOwnPropertyNames(
@@ -125,10 +128,16 @@ export default class RegionsPlugin {
         });
         this.wavesurfer.Region = Region;
 
+        // By default, scroll the container if the user drags a region
+        // within 5% of its edge
+        const scrollWidthProportion = 0.05;
         this._onBackendCreated = () => {
             this.wrapper = this.wavesurfer.drawer.wrapper;
+            this.orientation = this.wavesurfer.drawer.orientation;
             if (this.params.regions) {
                 this.params.regions.forEach(region => {
+                    region.edgeScrollWidth = this.params.edgeScrollWidth ||
+                        this.wrapper.clientWidth * scrollWidthProportion;
                     this.add(region);
                 });
             }
@@ -138,6 +147,7 @@ export default class RegionsPlugin {
         this.list = {};
         this._onReady = () => {
             this.wrapper = this.wavesurfer.drawer.wrapper;
+            this.vertical = this.wavesurfer.drawer.params.vertical;
             if (this.params.dragSelection) {
                 this.enableDragSelection(this.params);
             }
@@ -183,7 +193,13 @@ export default class RegionsPlugin {
      * @return {Region} The created region
      */
     add(params) {
-        if (this.wouldExceedMaxRegions()) return null;
+        if (this.wouldExceedMaxRegions()) {
+            return null;
+        }
+
+        if (!params.minLength && this.regionsMinLength) {
+            params = {...params, minLength: this.regionsMinLength};
+        }
 
         const region = new this.wavesurfer.Region(params, this.util, this.wavesurfer);
 
@@ -261,8 +277,12 @@ export default class RegionsPlugin {
             touchId = e.targetTouches ? e.targetTouches[0].identifier : null;
 
             // Store for scroll calculations
-            maxScroll = this.wrapper.scrollWidth - this.wrapper.clientWidth;
-            wrapperRect = this.wrapper.getBoundingClientRect();
+            maxScroll = this.wrapper.scrollWidth -
+                this.wrapper.clientWidth;
+            wrapperRect = this.util.withOrientation(
+                this.wrapper.getBoundingClientRect(),
+                this.vertical
+            );
 
             drag = true;
             start = this.wavesurfer.drawer.handleEvent(e, true);
@@ -293,6 +313,7 @@ export default class RegionsPlugin {
 
             region = null;
         };
+        this.wrapper.addEventListener('mouseleave', eventUp);
         this.wrapper.addEventListener('mouseup', eventUp);
         this.wrapper.addEventListener('touchend', eventUp);
 
@@ -303,9 +324,10 @@ export default class RegionsPlugin {
             document.body.removeEventListener('touchend', eventUp);
             this.wrapper.removeEventListener('touchend', eventUp);
             this.wrapper.removeEventListener('mouseup', eventUp);
+            this.wrapper.removeEventListener('mouseleave', eventUp);
         });
 
-        const eventMove = e => {
+        const eventMove = event => {
             if (!drag) {
                 return;
             }
@@ -313,20 +335,22 @@ export default class RegionsPlugin {
                 return;
             }
 
-            if (e.touches && e.touches.length > 1) {
+            if (event.touches && event.touches.length > 1) {
                 return;
             }
-            if (e.targetTouches && e.targetTouches[0].identifier != touchId) {
+            if (event.targetTouches && event.targetTouches[0].identifier != touchId) {
                 return;
             }
 
             // auto-create a region during mouse drag, unless region-count would exceed "maxRegions"
             if (!region) {
                 region = this.add(params || {});
-                if (!region) return;
+                if (!region) {
+                    return;
+                }
             }
 
-            const end = this.wavesurfer.drawer.handleEvent(e);
+            const end = this.wavesurfer.drawer.handleEvent(event);
             const startUpdate = this.wavesurfer.regions.util.getRegionSnapToGridValue(
                 start * duration
             );
@@ -338,10 +362,12 @@ export default class RegionsPlugin {
                 end: Math.max(endUpdate, startUpdate)
             });
 
+            let orientedEvent = this.util.withOrientation(event, this.vertical);
+
             // If scrolling is enabled
             if (scroll && container.clientWidth < this.wrapper.scrollWidth) {
                 // Check threshold based on mouse
-                const x = e.clientX - wrapperRect.left;
+                const x = orientedEvent.clientX - wrapperRect.left;
                 if (x <= scrollThreshold) {
                     scrollDirection = -1;
                 } else if (x >= wrapperRect.right - scrollThreshold) {
@@ -349,7 +375,7 @@ export default class RegionsPlugin {
                 } else {
                     scrollDirection = null;
                 }
-                scrollDirection && edgeScroll(e);
+                scrollDirection && edgeScroll(event);
             }
         };
         this.wrapper.addEventListener('mousemove', eventMove);
@@ -357,6 +383,12 @@ export default class RegionsPlugin {
         this.on('disable-drag-selection', () => {
             this.wrapper.removeEventListener('touchmove', eventMove);
             this.wrapper.removeEventListener('mousemove', eventMove);
+        });
+
+        this.wavesurfer.on('region-created', region => {
+            if (this.regionsMinLength) {
+                region.minLength = this.regionsMinLength;
+            }
         });
     }