+++ src/wavesurfer.js
@@ -114,6 +114,8 @@ import MediaElementWebAudio from './mediaelement-webaudio';
  * skipForward() and skipBackward() methods.
  * @property {boolean} splitChannels=false Render with separate waveforms for
  * the channels of the audio
+ * @property {SplitChannelOptions} splitChannelsOptions={} Options for splitChannel rendering
+ * @property {boolean} vertical=false Render the waveform vertically instead of horizontally.
  * @property {string} waveColor='#999' The fill color of the waveform after the
  * cursor.
  * @property {object} xhr={} XHR options. For example:
@@ -124,7 +126,7 @@ import MediaElementWebAudio from './mediaelement-webaudio';
  *     credentials: 'same-origin',
  *     redirect: 'follow',
  *     referrer: 'client',
- *     headers: [
+ *     requestHeaders: [
  *         {
  *             key: 'Authorization',
  *             value: 'my-token'
@@ -147,6 +149,28 @@ import MediaElementWebAudio from './mediaelement-webaudio';
  * passed to the plugin class constructor function
  * @property {PluginClass} instance The plugin instance factory, is called with
  * the dependency specified in extends. Returns the plugin class.
+ */
+
+/**
+ * @typedef {Object} SplitChannelOptions
+ * @desc parameters applied when splitChannels option is true
+ * @property {boolean} overlay=false determines whether channels are rendered on top of each other or on separate tracks
+ * @property {object} channelColors={} object describing color for each channel. Example:
+ * {
+ *     0: {
+ *         progressColor: 'green',
+ *         waveColor: 'pink'
+ *     },
+ *     1: {
+ *         progressColor: 'orange',
+ *         waveColor: 'purple'
+ *     }
+ * }
+ * @property {number[]} filterChannels=[] indexes of channels to be hidden from rendering
+ * @property {boolean} relativeNormalization=false determines whether
+ * normalization is done per channel or maintains proportionality between
+ * channels. Only applied when normalize and splitChannels are both true.
+ * @since 4.3.0
  */
 
 /**
@@ -267,8 +291,10 @@ export default class WaveSurfer extends util.Observer {
         splitChannelsOptions: {
             overlay: false,
             channelColors: {},
-            filterChannels: []
+            filterChannels: [],
+            relativeNormalization: false
         },
+        vertical: false,
         waveColor: '#999',
         xhr: {}
     };
@@ -338,7 +364,11 @@ export default class WaveSurfer extends util.Observer {
          * @private
          */
         this.params = Object.assign({}, this.defaultParams, params);
-
+        this.params.splitChannelsOptions = Object.assign(
+            {},
+            this.defaultParams.splitChannelsOptions,
+            params.splitChannelsOptions
+        );
         /** @private */
         this.container =
             'string' == typeof params.container
@@ -373,7 +403,11 @@ export default class WaveSurfer extends util.Observer {
         }
 
         if (this.params.rtl === true) {
-            util.style(this.container, { transform: 'rotateY(180deg)' });
+            if (this.params.vertical === true) {
+                util.style(this.container, { transform: 'rotateX(180deg)' });
+            } else {
+                util.style(this.container, { transform: 'rotateY(180deg)' });
+            }
         }
 
         if (this.params.backgroundColor) {
@@ -900,20 +934,23 @@ export default class WaveSurfer extends util.Observer {
         }
         this.fireEvent('interaction', () => this.seekTo(progress));
 
+        const isWebAudioBackend = this.params.backend === 'WebAudio';
         const paused = this.backend.isPaused();
-        // avoid draw wrong position while playing backward seeking
-        if (!paused) {
+
+        if (isWebAudioBackend && !paused) {
             this.backend.pause();
         }
+
         // avoid small scrolls while paused seeking
         const oldScrollParent = this.params.scrollParent;
         this.params.scrollParent = false;
         this.backend.seekTo(progress * this.getDuration());
         this.drawer.progress(progress);
 
-        if (!paused) {
+        if (isWebAudioBackend && !paused) {
             this.backend.play();
         }
+
         this.params.scrollParent = oldScrollParent;
         this.fireEvent('seek', progress);
     }
@@ -1386,6 +1423,12 @@ export default class WaveSurfer extends util.Observer {
             }
         }
 
+        // loadBuffer(url, peaks, duration) requires that url is a string
+        // but users can pass in a HTMLMediaElement to WaveSurfer
+        if (this.params.backend === 'WebAudio' && url instanceof HTMLMediaElement) {
+            url = url.src;
+        }
+
         switch (this.params.backend) {
             case 'WebAudio':
                 return this.loadBuffer(url, peaks, duration);
@@ -1399,6 +1442,7 @@ export default class WaveSurfer extends util.Observer {
      * Loads audio using Web Audio buffer backend.
      *
      * @private
+     * @emits WaveSurfer#waveform-ready
      * @param {string} url URL of audio file
      * @param {number[]|Number.<Array[]>} peaks Peaks data
      * @param {?number} duration Optional duration of audio file
@@ -1415,6 +1459,7 @@ export default class WaveSurfer extends util.Observer {
         if (peaks) {
             this.backend.setPeaks(peaks, duration);
             this.drawBuffer();
+            this.fireEvent('waveform-ready');
             this.tmpEvents.push(this.once('interaction', load));
         } else {
             return load();
@@ -1425,6 +1470,7 @@ export default class WaveSurfer extends util.Observer {
      * Either create a media element, or load an existing media element.
      *
      * @private
+     * @emits WaveSurfer#waveform-ready
      * @param {string|HTMLMediaElement} urlOrElt Either a path to a media file, or an
      * existing HTML5 Audio/Video Element
      * @param {number[]|Number.<Array[]>} peaks Array of peaks. Required to bypass web audio
@@ -1459,9 +1505,11 @@ export default class WaveSurfer extends util.Observer {
             this.backend.once('error', err => this.fireEvent('error', err))
         );
 
+        // If peaks are provided, render them and fire the `waveform-ready` event.
         if (peaks) {
             this.backend.setPeaks(peaks, duration);
             this.drawBuffer();
+            this.fireEvent('waveform-ready');
         }
 
         // If no pre-decoded peaks are provided, or are provided with
@@ -1562,7 +1610,8 @@ export default class WaveSurfer extends util.Observer {
     }
 
     /**
-     * Exports PCM data into a JSON array and opens in a new window.
+     * Exports PCM data into a JSON array and optionally opens in a new window
+     * as valid JSON Blob instance.
      *
      * @param {number} length=1024 The scale in which to export the peaks
      * @param {number} accuracy=10000
@@ -1582,16 +1631,18 @@ export default class WaveSurfer extends util.Observer {
             peaks,
             val => Math.round(val * accuracy) / accuracy
         );
-        return new Promise((resolve, reject) => {
-            const json = JSON.stringify(arr);
 
-            if (!noWindow) {
-                window.open(
-                    'data:application/json;charset=utf-8,' +
-                        encodeURIComponent(json)
+        return new Promise((resolve, reject) => {
+            if (!noWindow){
+                const blobJSON = new Blob(
+                    [JSON.stringify(arr)],
+                    {type: 'application/json;charset=utf-8'}
                 );
+                const objURL = URL.createObjectURL(blobJSON);
+                window.open(objURL);
+                URL.revokeObjectURL(objURL);
             }
-            resolve(json);
+            resolve(arr);
         });
     }
 
@@ -1632,6 +1683,14 @@ export default class WaveSurfer extends util.Observer {
      */
     cancelAjax() {
         if (this.currentRequest && this.currentRequest.controller) {
+            // If the current request has a ProgressHandler, then its ReadableStream might need to be cancelled too
+            // See: Wavesurfer issue #2042
+            // See Firefox bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1583815
+            if (this.currentRequest._reader) {
+                // Ignoring exceptions thrown by call to cancel()
+                this.currentRequest._reader.cancel().catch(err => {});
+            }
+
             this.currentRequest.controller.abort();
             this.currentRequest = null;
         }
@@ -1683,6 +1742,8 @@ export default class WaveSurfer extends util.Observer {
         }
         if (this.backend) {
             this.backend.destroy();
+            // clears memory usage
+            this.backend = null;
         }
         if (this.drawer) {
             this.drawer.destroy();